Make nullable textknob single widget
