+++ packages/widgetbook/test/src/knobs/nullable_text_knob_test.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:widgetbook/src/knobs/knobs.dart';
-import 'package:widgetbook/src/knobs/nullable_text_knob.dart';
+import 'package:widgetbook/src/knobs/text_knob.dart';
 
 import '../../helper/widget_test_helper.dart';
 import 'knobs_test.dart';
